+++ src/org/joml/Matrix4d.java
@@ -376,13 +376,13 @@ public class Matrix4d implements Externalizable {
     }
 
     /**
-     * Set this matrix to be equivalent to the rotation specified by the given {@link AxisAngle4f}.
+     * Set this matrix to be equivalent to the rotation specified by the given {@link AngleAxis4f}.
      * 
      * @param axisAngle
-     *          the {@link AxisAngle4f}
+     *          the {@link AngleAxis4f}
      * @return this
      */
-    public Matrix4d set(AxisAngle4f axisAngle) {
+    public Matrix4d set(AngleAxis4f axisAngle) {
         double x = axisAngle.x;
         double y = axisAngle.y;
         double z = axisAngle.z;
@@ -1223,15 +1223,15 @@ public class Matrix4d implements Externalizable {
 
     /**
      * Get the rotational component of <code>this</code> matrix and store the represented rotation
-     * into the given {@link AxisAngle4f}.
+     * into the given {@link AngleAxis4f}.
      * 
-     * @see AxisAngle4f#set(Matrix4d)
+     * @see AngleAxis4f#set(Matrix4d)
      * 
      * @param dest
-     *          the destination {@link AxisAngle4f}
+     *          the destination {@link AngleAxis4f}
      * @return this
      */
-    public Matrix4d get(AxisAngle4f dest) {
+    public Matrix4d get(AngleAxis4f dest) {
         dest.set(this);
         return this;
     }
@@ -2200,23 +2200,23 @@ public class Matrix4d implements Externalizable {
     }
 
     /**
-     * Set this matrix to a rotation transformation using the given {@link AxisAngle4f}.
+     * Set this matrix to a rotation transformation using the given {@link AngleAxis4f}.
      * <p>
      * The resulting matrix can be multiplied against another transformation
      * matrix to obtain an additional rotation.
      * <p>
      * In order to apply the rotation transformation to an existing transformation,
-     * use {@link #rotate(AxisAngle4f) rotate()} instead.
+     * use {@link #rotate(AngleAxis4f) rotate()} instead.
      * <p>
      * Reference: <a href="http://en.wikipedia.org/wiki/Rotation_matrix#Axis_and_angle">http://en.wikipedia.org</a>
      *
-     * @see #rotate(AxisAngle4f)
+     * @see #rotate(AngleAxis4f)
      * 
      * @param angleAxis
-     *          the {@link AxisAngle4f} (needs to be {@link AxisAngle4f#normalize() normalized})
+     *          the {@link AngleAxis4f} (needs to be {@link AngleAxis4f#normalize() normalized})
      * @return this
      */
-    public Matrix4d rotation(AxisAngle4f angleAxis) {
+    public Matrix4d rotation(AngleAxis4f angleAxis) {
         return rotation(angleAxis.angle, angleAxis.x, angleAxis.y, angleAxis.z);
     }
 
@@ -2509,52 +2509,52 @@ public class Matrix4d implements Externalizable {
     }
 
     /**
-     * Apply a rotation transformation, rotating about the given {@link AxisAngle4f}, to this matrix.
+     * Apply a rotation transformation, rotating about the given {@link AngleAxis4f}, to this matrix.
      * <p>
-     * If <code>M</code> is <code>this</code> matrix and <code>A</code> the rotation matrix obtained from the given {@link AxisAngle4f},
+     * If <code>M</code> is <code>this</code> matrix and <code>A</code> the rotation matrix obtained from the given {@link AngleAxis4f},
      * then the new matrix will be <code>M * A</code>. So when transforming a
      * vector <code>v</code> with the new matrix by using <code>M * A * v</code>,
-     * the {@link AxisAngle4f} rotation will be applied first!
+     * the {@link AngleAxis4f} rotation will be applied first!
      * <p>
      * In order to set the matrix to a rotation transformation without post-multiplying,
-     * use {@link #rotation(AxisAngle4f)}.
+     * use {@link #rotation(AngleAxis4f)}.
      * <p>
      * Reference: <a href="http://en.wikipedia.org/wiki/Rotation_matrix#Axis_and_angle">http://en.wikipedia.org</a>
      * 
      * @see #rotate(double, double, double, double)
-     * @see #rotation(AxisAngle4f)
+     * @see #rotation(AngleAxis4f)
      * 
      * @param axisAngle
-     *          the {@link AxisAngle4f} (needs to be {@link AxisAngle4f#normalize() normalized})
+     *          the {@link AngleAxis4f} (needs to be {@link AngleAxis4f#normalize() normalized})
      * @return this
      */
-    public Matrix4d rotate(AxisAngle4f axisAngle) {
+    public Matrix4d rotate(AngleAxis4f axisAngle) {
         return rotate(axisAngle.angle, axisAngle.x, axisAngle.y, axisAngle.z);
     }
 
     /**
-     * Apply a rotation transformation, rotating about the given {@link AxisAngle4f} and store the result in <code>dest</code>.
+     * Apply a rotation transformation, rotating about the given {@link AngleAxis4f} and store the result in <code>dest</code>.
      * <p>
-     * If <code>M</code> is <code>this</code> matrix and <code>A</code> the rotation matrix obtained from the given {@link AxisAngle4f},
+     * If <code>M</code> is <code>this</code> matrix and <code>A</code> the rotation matrix obtained from the given {@link AngleAxis4f},
      * then the new matrix will be <code>M * A</code>. So when transforming a
      * vector <code>v</code> with the new matrix by using <code>M * A * v</code>,
-     * the {@link AxisAngle4f} rotation will be applied first!
+     * the {@link AngleAxis4f} rotation will be applied first!
      * <p>
      * In order to set the matrix to a rotation transformation without post-multiplying,
-     * use {@link #rotation(AxisAngle4f)}.
+     * use {@link #rotation(AngleAxis4f)}.
      * <p>
      * Reference: <a href="http://en.wikipedia.org/wiki/Rotation_matrix#Axis_and_angle">http://en.wikipedia.org</a>
      * 
      * @see #rotate(double, double, double, double)
-     * @see #rotation(AxisAngle4f)
+     * @see #rotation(AngleAxis4f)
      * 
      * @param axisAngle
-     *          the {@link AxisAngle4f} (needs to be {@link AxisAngle4f#normalize() normalized})
+     *          the {@link AngleAxis4f} (needs to be {@link AngleAxis4f#normalize() normalized})
      * @param dest
      *          will hold the result
      * @return this
      */
-    public Matrix4d rotate(AxisAngle4f axisAngle, Matrix4d dest) {
+    public Matrix4d rotate(AngleAxis4f axisAngle, Matrix4d dest) {
         return rotate(axisAngle.angle, axisAngle.x, axisAngle.y, axisAngle.z, dest);
     }
 
@@ -5075,16 +5075,16 @@ public class Matrix4d implements Externalizable {
     }
 
     /**
-     * Return the specified {@link AxisAngle4f}.
+     * Return the specified {@link AngleAxis4f}.
      * <p>
      * When using method chaining in a fluent interface style, this method can be used to switch
-     * the <i>context object</i>, on which further method invocations operate, to be the given {@link AxisAngle4f}.
+     * the <i>context object</i>, on which further method invocations operate, to be the given {@link AngleAxis4f}.
      * 
      * @param a
-     *          the {@link AxisAngle4f} to return
+     *          the {@link AngleAxis4f} to return
      * @return that quaternion
      */
-    public AxisAngle4f with(AxisAngle4f a) {
+    public AngleAxis4f with(AngleAxis4f a) {
         return a;
     }
